pathology/bazel_requirements.txt
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
# ==============================================================================
# Bazel requirements
absl_py==2.1.0
brotli==1.1.0
flask_cors==5.0.0
gunicorn==23.0.0
flask_compress==1.17
imagecodecs==2024.12.30
dnspython==2.7.0
docopt==0.6.2
rpds-py==0.22.3
attrs==24.2.0
referencing==0.35.1
jsonschema-specifications==2024.10.1
regex==2024.11.6
pyarrow-hotfix==0.6
pydot==3.0.3
objsize==0.7.0
pyarrow==18.1.0
orjson==3.10.12
sortedcontainers==2.4.0
zstandard==0.23.0
jsonschema==4.23.0
fasteners==0.19
hdfs==2.7.3
crcmod==1.7
fastavro==1.9.7
dill==0.3.9
jsonpickle==4.0.1
cloudpickle==3.1.0
pymongo==4.10.1
apache-beam==2.61.0
pytz==2024.2
tzdata==2024.2
uritemplate==4.1.1
six==1.17.0
python-dateutil==2.9.0.post0
mypy-extensions==1.0.0
packaging==24.2
typing-inspect==0.9.0
marshmallow==3.23.1
dataclasses-json==0.6.7
redis==5.2.1
PyYAML==6.0.2
google-cloud-bigquery==3.27.0
tifffile==2024.12.12
pillow==11.1.0
openslide-python==1.4.1
numpy==2.0.2
opencv-python-headless==4.10.0.84
google-api-python-client==2.156.0
pandas==2.2.3
typing-extensions==4.12.2
opentelemetry-semantic-conventions==0.50b0
grpcio-status==1.68.1
opentelemetry-sdk==1.29.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
google-cloud-storage==2.19.0
google-cloud-pubsub==2.27.1
pyparsing==3.2.0
httplib2==0.22.0
requests-toolbelt==1.0.0
google-auth-httplib2==0.2.0
requests_mock==1.12.1
grpc-interceptor==0.15.4
sqlparse==0.5.3
google-cloud-spanner==3.51.0
MarkupSafe==3.0.2
click==8.1.7
Werkzeug==3.1.3
blinker==1.9.0
Jinja2==3.1.4
itsdangerous==2.2.0
flask==3.1.0
flask==3.1.0
pydicom==3.0.1
zipp==3.21.0
wrapt==1.17.0
deprecated==1.2.15
importlib-metadata==8.5.0
google-cloud-audit-log==0.3.0
google_cloud_appengine_logging==1.5.0
opentelemetry-api==1.29.0
google_cloud_core==2.4.1
google_cloud_logging==3.11.3
psutil==6.1.1
absl_py==2.1.0
google-api-core==2.24.0
requests==2.32.3
google-auth==2.37.0
googleapis-common-protos==1.66.0
protobuf==5.29.1
proto-plus==1.25.0
urllib3==2.3.0
charset-normalizer==3.4.0
idna==3.10
certifi==2024.12.14
rsa==4.9
pyasn1-modules==0.4.1
cachetools==5.5.0
pyasn1==0.6.1
google-cloud-secret-manager==2.22.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1